+++ test/it/jokeutils/banner_test.sh
@@ -1,0 +1,3 @@
+TestBanner() {
+    banner HI | wc -l | tr -d ' '
+}