+++ test/it/jokeutils/cmatrix_test.sh
@@ -1,0 +1,4 @@
+TestCmatrixNoTTY() {
+    cmatrix
+    echo "rc:$?"
+}