+++ test/it/jokeutils/cowthink_test.sh
@@ -1,0 +1,3 @@
+TestCowthink() {
+    cowthink hi | head -n 4 | tail -n 1
+}