+++ test/it/spec/fortune_spec.sh
@@ -1,0 +1,8 @@
+Describe 'fortune'
+    Include jokeutils/fortune_test.sh
+    It 'prints a single adage line'
+        When call TestFortune
+        The output should equal '1'
+        The status should be success
+    End
+End